requirements.txt
glfw
imgui[glfw]
Pillow
miniaudio
freetype-py
moderngl
numpy
numba
